pyproject.toml
[project]
name = "performance-rrt"
version = "0.1.0"
description = "Various RRT* implementations for educational purposes"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "line-profiler>=4.2.0",
    "matplotlib>=3.9.4",
    "numba>=0.60.0",
    "numpy>=2.0.2",
    "pybind11>=2.13.6",
    "rerun-sdk>=0.22.1",
    "scipy>=1.13.1",
    "snakeviz>=2.2.2",
]

[build-system]
requires = ["setuptools>=40.8.0", "wheel", "pybind11"]
build-backend = "setuptools.build_meta"
